+++ CorrugatedIron/RiakClient.cs
@@ -697,6 +697,32 @@ namespace CorrugatedIron
                 return result;
             }
         }
+
+        /// <summary>
+        /// Reset the properties on a bucket back to their defaults.
+        /// </summary>
+        /// <param name="bucket">The name of the bucket to reset the properties on.</param>
+        /// <returns>An indication of success or failure.</returns>
+        /// <remarks>This function requires Riak v1.3+.</remarks>
+        public RiakResult ResetBucketProperties(string bucket)
+        {
+            var request = new RiakRestRequest(ToBucketPropsUri(bucket), RiakConstants.Rest.HttpMethod.Delete);
+
+            var result = UseConnection(conn => conn.RestRequest(request));
+            if(result.IsSuccess)
+            {
+                switch (result.Value.StatusCode)
+                {
+                    case HttpStatusCode.NoContent:
+                        return result;
+                    case HttpStatusCode.NotFound:
+                        return RiakResult.Error(ResultCode.NotFound, "Bucket {0} not found.".Fmt(bucket), false);
+                    default:
+                        return RiakResult.Error(ResultCode.InvalidResponse, "Unexpected Status Code: {0} ({1})".Fmt(result.Value.StatusCode, (int)result.Value.StatusCode), result.NodeOffline);
+                }
+            }
+            return result;
+        }
         
         /// <summary>
         /// Get the results of an index query prepared for use in a <see cref="CorrugatedIron.Models.MapReduce.RiakMapReduceQuery"/>
@@ -911,6 +937,11 @@ namespace CorrugatedIron
         private static string ToBucketUri(string bucket)
         {
             return "{0}/{1}".Fmt(RiakConstants.Rest.Uri.RiakRoot, HttpUtility.UrlEncode(bucket));
+        }
+
+        private static string ToBucketPropsUri(string bucket)
+        {
+            return RiakConstants.Rest.Uri.BucketPropsFmt.Fmt(HttpUtility.UrlEncode(bucket));
         }
     }
 }