+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
 Release Notes
 =============
 
+### Fixes
+
+* [Issue 94](https://github.com/DistributedNonsense/CorrugatedIron/issues/94) - Riak 1.3 Reset Bucket Properties functionality
+
 v1.2.0
 ------